requirements.txt
matplotlib >=3.7.0
numpy >=1.24.0
onnx >=1.14.0
opencv-python >=4.4.0.0
Pillow>=9.5.0
scipy
torch
torchvision
rich